Add messages visualization controller and health checks.

+++ java_task_6/micro-collector/src/main/java/com/epam/jmp/microcollector/repositories/MessagesRepository.java
@@ -4,6 +4,8 @@ import com.epam.jmp.microcollector.model.Message;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
+
 @Repository
-public interface MessagesRepository extends MongoRepository<Message, Long> {
+public interface MessagesRepository extends MongoRepository<Message, BigInteger> {
 }
